+++ src/components/TodoInfo/TodoInfo.tsx
@@ -1,99 +1,27 @@
-import React, { useState } from 'react';
-import { User } from '../../types/User';
 import { Todo } from '../../types/Todo';
-import users from '../../api/users';
-import { getUserById } from '../../services/user';
+import classNames from 'classnames';
+import { UserInfo } from '../UserInfo';
 
 type Props = {
-  onSubmit: (todo: Todo) => void;
   todos: Todo[];
 };
 
-export const TodoInfo: React.FC<Props> = ({ onSubmit, todos }) => {
-  const [title, setTitle] = useState('');
-  const [hasTitleError, setHasTitleError] = useState(false);
-
-  const [userId, setUserId] = useState(0);
-  const [hasUserIdError, setHasUserIdError] = useState(false);
-
-  const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const rawValue = e.target.value;
-
-    const cleanedValue = rawValue.replace(/[^a-zA-Zа-яА-ЯґҐєЄіІїЇ0-9 ]/g, '');
-
-    setTitle(cleanedValue);
-    setHasTitleError(false);
-  };
-
-  const handleUserIdChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    setUserId(+e.target.value);
-    setHasUserIdError(false);
-  };
-
-  const maxId = todos.length ? Math.max(...todos.map(todo => todo.id)) : 0;
-
-  const handleSubmit = (event: React.FormEvent) => {
-    event.preventDefault();
-
-    setHasTitleError(!title);
-    setHasUserIdError(!userId);
-
-    if (!title || !userId) {
-      return;
-    }
-
-    onSubmit({
-      id: maxId + 1,
-      title,
-      completed: false,
-      userId,
-      user: getUserById(userId),
-    });
-
-    setTitle('');
-    setUserId(0);
-  };
-
+export const TodoInfo: React.FC<Props> = ({ todos }) => {
   return (
-    <form action="/api/todos" method="POST" onSubmit={handleSubmit}>
-      <label className="field">
-        Title:
-        <input
-          type="text"
-          data-cy="titleInput"
-          placeholder="Enter a title"
-          value={title}
-          onChange={handleTitleChange}
-        />
-        {hasTitleError && <span className="error">Please enter a title</span>}
-      </label>
-
-      <div className="field">
-        <label htmlFor="user-id">User:</label>
-        <select
-          id="user-id"
-          data-cy="userSelect"
-          required
-          value={userId}
-          onChange={handleUserIdChange}
+    <>
+      {todos.map(todo => (
+        <article
+          key={todo.id}
+          data-id={todo.id}
+          className={classNames('TodoInfo', {
+            'TodoInfo--completed': todo.completed,
+          })}
         >
-          <option value="0" disabled>
-            Choose a user
-          </option>
+          <h2 className="TodoInfo__title">{todo.title}</h2>
 
-          {users.map((user: User) => (
-            <option value={user.id} key={user.id}>
-              {user.name}
-            </option>
-          ))}
-        </select>
-
-        {hasUserIdError && <span className="error">Please choose a user</span>}
-      </div>
-
-      <button type="submit" data-cy="submitButton">
-        Add
-      </button>
-    </form>
+          <UserInfo user={todo.user!} />
+        </article>
+      ))}
+    </>
   );
 };
