+++ src/components/TodoList/TodoList.tsx
@@ -1,6 +1,5 @@
-import classNames from 'classnames';
 import { Todo } from '../../types/Todo';
-import { UserInfo } from '../UserInfo';
+import { TodoInfo } from '../TodoInfo';
 
 type Props = {
   todos: Todo[];
@@ -9,19 +8,7 @@ type Props = {
 export const TodoList: React.FC<Props> = ({ todos }) => {
   return (
     <section className="TodoList">
-      {todos.map(todo => (
-        <article
-          key={todo.id}
-          data-id={todo.id}
-          className={classNames('TodoInfo', {
-            'TodoInfo--completed': todo.completed,
-          })}
-        >
-          <h2 className="TodoInfo__title">{todo.title}</h2>
-
-          <UserInfo user={todo.user!} />
-        </article>
-      ))}
+      <TodoInfo todos={todos} />
     </section>
   );
 };
